Added more grammar, fixed > ambiguity
Array accesses, ranges

Fixed ambiguity with assignment - '=' was recognised as a binary op,
so was being consumed into expression. Instead, a separate production
was added just for assignments.

Fixed ambiguity with templates and the '>' symbol. Bit
of a hack involving passing booleans around, but it's
and optional parameter, so it's not too serious.

Added comptime types and variables